# api/openapi.yaml
openapi: 3.0.3
info:
  title: Ted control plane API
  version: 1.0.0
  description: Local unauthenticated control plane. Bind to loopback unless protected
    by a trusted proxy. All request objects reject unknown properties. IDs and cursors
    are opaque runtime identifiers; cursors are nonnegative integers. Bodies limited
    to 2 MiB. JSON errors share one envelope. No wildcard CORS.
paths:
...
  /v1/projects:
    get:
      operationId: ListProjects
      description: ListProjects
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '400': *id001
        '404': *id002
        '405': *id003
        '409': *id004
        '410': *id005
        '413': *id006
        '415': *id007
        '500': *id008
        '503': *id009
    post:
      operationId: CreateProject
      description: CreateProject
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400': *id001
        '404': *id002
        '405': *id003
        '409': *id004
        '410': *id005
        '413': *id006
        '415': *id007
        '500': *id008
        '503': *id009
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
  /v1/projects/{project_id}:
    get:
      operationId: GetProject
      description: GetProject
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400': *id001
        '404': *id002
        '405': *id003
        '409': *id004
        '410': *id005
        '413': *id006
        '415': *id007
        '500': *id008
        '503': *id009
      parameters:
      - name: project_id
        in: path
        required: true
        schema: &id010
          type: string
          minLength: 1
          maxLength: 256
    patch:
      operationId: PatchProject
      description: Update project name or defaults; defaults affect future agents
        only.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400': *id001
        '404': *id002
        '405': *id003
        '409': *id004
        '410': *id005
        '413': *id006
        '415': *id007
        '500': *id008
        '503': *id009
      parameters:
      - name: project_id
        in: path
        required: true
        schema: *id010
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchProjectRequest'
    delete:
      operationId: DeleteProject
      description: Delete a project and all its settled agents and chat history. Returns 409 while any agents are unsettled or still stopping, or have children in other projects. The project directory is not removed.
      responses:
        '204':
          description: Success
        '400': *id001
        '404': *id002
        '405': *id003
        '409': *id004
        '410': *id005
        '413': *id006
        '415': *id007
        '500': *id008
        '503': *id009
      parameters:
      - name: project_id
        in: path
        required: true
        schema: *id010
  /v1/projects/{project_id}/branches:
    get:
      operationId: ListProjectBranches
      description: List remote-qualified Git branches already known in the local checkout. No network fetch is performed.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectBranches'
        '400': *id001
        '404': *id002
        '405': *id003
        '409': *id004
        '410': *id005
        '413': *id006
        '415': *id007
        '500': *id008
        '503': *id009
      parameters:
      - name: project_id
        in: path
        required: true
        schema: *id010
  /v1/agents:
    get:
      operationId: ListAgents
      description: >-
        List agents. Without page or page_size, returns the legacy full array and
        ordering. When either pagination parameter is supplied, an omitted page
        defaults to 1 and an omitted page_size defaults to 25. Paginated results
        are sorted by updated_at descending, then id ascending, and X-Total-Count
        reports the matching total after filtering.
      responses:
        '200':
          description: Success
          headers:
            X-Total-Count:
              description: Total number of agents matching the include_settled and project_id filters when pagination parameters are supplied.
              schema:
                type: integer
                format: int64
                minimum: 0
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Agent'
        '400': *id001
        '404': *id002
        '405': *id003
        '409': *id004
        '410': *id005
        '413': *id006
        '415': *id007
        '500': *id008
        '503': *id009
      parameters:
      - name: include_settled
        in: query
        required: false
        schema:
          type: boolean
          default: false
      - name: project_id
        in: query
        required: false
        schema: *id010
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int64
          minimum: 1
          default: 1
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          format: int64
          minimum: 1
          default: 25
    post:
      operationId: CreateAgent
      description: Create an agent; optional prompt queues its first turn. Optional
        Idempotency-Key makes retries safe; differing payload with the same key returns
        409.
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400': *id001
        '404': *id002
        '405': *id003
        '409': *id004
        '410': *id005
        '413': *id006
        '415': *id007
        '500': *id008
        '503': *id009
      parameters:
      - &id011
        name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
          minLength: 1
          maxLength: 256
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
  /v1/agents/{agent_id}:
    get:
      operationId: GetAgent
      description: GetAgent
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400': *id001
        '404': *id002
        '405': *id003
        '409': *id004
        '410': *id005
        '413': *id006
        '415': *id007
        '500': *id008
        '503': *id009
      parameters:
      - name: agent_id
        in: path
        required: true
        schema: *id010
    patch:
      operationId: PatchAgent
      description: >-
        Update settled state or advance the durable shared read cursor to exactly
        the supplied event cursor. Setting `settled` to true atomically settles
        all descendants, including across projects and through already-settled
        agents; queued work remains held and running turns are cancelled after
        persistence. Setting it to false restores only the requested agent and
        leaves queued work held. Exactly one property is required. Lower or
        duplicate read cursors are no-ops; cursors beyond the current agent event
        stream are rejected.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400': *id001
        '404': *id002
        '405': *id003
        '409': *id004
        '410': *id005
        '413': *id006
        '415': *id007
        '500': *id008
        '503': *id009
      parameters:
      - name: agent_id
        in: path
        required: true
        schema: *id010
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchAgentRequest'
  /v1/agents/{agent_id}/pull-request:
    get:
      operationId: GetAgentPullRequest
      description: Resolve the GitHub pull request associated with the agent's current recorded branch. Lookup failures and misses return the branch without a number.
      responses:
        '200':
          description: Branch association. Empty when the agent has no eligible branch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentPullRequest'
        '400': *id001
        '404': *id002
        '405': *id003
        '409': *id004
        '410': *id005
        '413': *id006
        '415': *id007
        '500': *id008
        '503': *id009
      parameters:
      - name: agent_id
        in: path
        required: true
        schema: *id010
  /v1/agents/{agent_id}/workspace:
    patch:
      operationId: PatchAgentWorkspace
      description: Change the workspace selection before the first nonempty message. The selection locks irrevocably on first send.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400': *id001
        '404': *id002
        '405': *id003
        '409': *id004
        '410': *id005
        '413': *id006
        '415': *id007
        '500': *id008
        '503': *id009
      parameters:
      - name: agent_id
        in: path
        required: true
        schema: *id010
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceSelection'
  /v1/agents/{agent_id}/settings:
    patch:
      operationId: PatchAgentSettings
      description: Change settings for future turns; the active turn keeps its captured
        settings.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400': *id001
        '404': *id002
        '405': *id003
        '409': *id004
        '410': *id005
        '413': *id006
        '415': *id007
        '500': *id008
        '503': *id009
      parameters:
      - name: agent_id
        in: path
        required: true
        schema: *id010
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingsPatch'
  /v1/agents/{agent_id}/messages:
    get:
      operationId: ListMessages
      description: Inspect the queue, including terminal entries.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QueuedMessage'
        '400': *id001
        '404': *id002
        '405': *id003
        '409': *id004
        '410': *id005
        '413': *id006
        '415': *id007
        '500': *id008
        '503': *id009
      parameters:
      - name: agent_id
        in: path
        required: true
        schema: *id010
    post:
      operationId: SubmitMessage
      description: Queue a message. Optional Idempotency-Key deduplicates submissions
        across HTTP and WebSocket; different text, kind, or sender with the same key
        returns 409. Omitted kind and explicit user kind are equivalent.
      responses:
        '202':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueuedMessage'
        '400': *id001
        '404': *id002
        '405': *id003
        '409': *id004
        '410': *id005
        '413': *id006
        '415': *id007
        '500': *id008
        '503': *id009
      parameters:
      - name: agent_id
        in: path
        required: true
        schema: *id010
      - *id011
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitMessageRequest'
  /v1/agents/{agent_id}/messages/{message_id}:
    delete:
      operationId: DeletePending
      description: DeletePending
      responses:
        '204':
          description: Success
        '400': *id001
        '404': *id002
        '405': *id003
        '409': *id004
        '410': *id005
        '413': *id006
        '415': *id007
        '500': *id008
        '503': *id009
      parameters:
      - name: agent_id
        in: path
        required: true
        schema: *id010
      - name: message_id
        in: path
        required: true
        schema: *id010
  /v1/agents/{agent_id}/stop:
    post:
      operationId: StopAgent
      description: Stop only the named turn (turn_id is its queued message ID); advance
        pending queue. Safe to retry without stopping a newer turn.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400': *id001
        '404': *id002
        '405': *id003
        '409': *id004
        '410': *id005
        '413': *id006
        '415': *id007
        '500': *id008
        '503': *id009
      parameters:
      - name: agent_id
        in: path
        required: true
        schema: *id010
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StopRequest'
  /v1/agents/{agent_id}/continue:
    post:
      operationId: ContinueAgent
      description: Release queue hold and schedule pending messages. No request body.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400': *id001
        '404': *id002
        '405': *id003
        '409': *id004
        '410': *id005
        '413': *id006
        '415': *id007
        '500': *id008
        '503': *id009
      parameters:
      - name: agent_id
        in: path
        required: true
        schema: *id010
  /v1/agents/{agent_id}/history:
    get:
      operationId: GetHistory
      description: Conversation messages after a zero-based consumed-message cursor.
        next_cursor equals after plus returned count. Ahead-of-history cursors return
        410; conversation compaction may invalidate a cursor.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryPage'
        '400': *id001
        '404': *id002
        '405': *id003
        '409': *id004
        '410': *id005
        '413': *id006
        '415': *id007
        '500': *id008
        '503': *id009
      parameters:
      - name: agent_id
        in: path
        required: true
        schema: *id010
      - &id012
        name: after
        in: query
        required: false
        schema:
          type: integer
          format: int64
          minimum: 0
          default: 0
      - &id013
        name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int64
          minimum: 1
          maximum: 1000
          default: 100
  /v1/agents/{agent_id}/events:
    get:
      operationId: GetEvents
      description: Lifetime event log, exclusive after cursor. next_cursor is last
        returned cursor or after on empty page. Invalid cursors return 410. Full oversized
        output is retained.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventPage'
        '400': *id001
        '404': *id002
        '405': *id003
        '409': *id004
        '410': *id005
        '413': *id006
        '415': *id007
        '500': *id008
        '503': *id009
      parameters:
      - name: agent_id
        in: path
        required: true
        schema: *id010
      - *id012
      - *id013
  /v1/agents/{agent_id}/events/{cursor}:
    get:
      operationId: GetEvent
      description: Fetch one complete retained event, including output referenced
        by an event_ref WebSocket frame. No output truncation.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400': *id001
        '404': *id002
        '405': *id003
        '409': *id004
        '410': *id005
        '413': *id006
        '415': *id007
        '500': *id008
        '503': *id009
      parameters:
      - name: agent_id
        in: path
        required: true
        schema: *id010
      - name: cursor
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
  /v1/models:
    get:
      operationId: ListModels
      description: ListModels
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Model'
        '400': *id001
        '404': *id002
        '405': *id003
        '409': *id004
        '410': *id005
        '413': *id006
        '415': *id007
        '500': *id008
        '503': *id009
...
components:
  schemas:
...
    Project:
      type: object
      additionalProperties: false
      properties:
        git_branch:
          type: string
          description: Live server-side Git branch (or detached commit), populated on GET project; omitted outside Git repositories.
        id: *id010
        name:
          type: string
        root:
          type: string
        defaults:
          $ref: '#/components/schemas/Settings'
        workspace_defaults:
          $ref: '#/components/schemas/WorkspaceSelection'
      required:
      - id
      - name
      - root
      - defaults
    CreateProjectRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 256
        root:
          type: string
          minLength: 1
          maxLength: 4096
        defaults:
          $ref: '#/components/schemas/Settings'
        workspace_defaults:
          $ref: '#/components/schemas/WorkspaceSelection'
      required:
      - name
      - root
      - defaults
    PatchProjectRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 256
        defaults:
          $ref: '#/components/schemas/Settings'
        workspace_defaults:
          $ref: '#/components/schemas/WorkspaceSelection'
      minProperties: 1
...